% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2forest_2.R
\name{forestplotlist_es}
\alias{forestplotlist_es}
\title{This function takes row of a standard data frame of FIA, FR or ES and creates a list following the format of a forest object}
\usage{
forestplotlist_es(
  id,
  version,
  country,
  tree,
  regen,
  shrub,
  filterNA,
  filterDead,
  minDBH,
  filterminDBH,
  setDefaults,
  .verbose = TRUE
)
}
\arguments{
\item{id}{id code for plot}

\item{version}{NA as default, enter "ifn2" "ifn3" OR "ifn4" only for SPAIN}

\item{country}{country code "FR" "ES" or "US"}

\item{tree}{standard data frame with tree info}

\item{regen}{standard data frame with regen info}

\item{filterNA}{if TRUE records with NA for DBH or HT  are eliminated}

\item{filterDead}{if TRUE only records of alive trees are conserved}

\item{minDBH}{min DBH value}

\item{filterminDBH}{if TRUE records are eliminated below minDBH}

\item{setDefaults}{set defaults params for root depth}

\item{year}{year of sampling , NA as default for ES}

\item{understory}{standard data frame with understory info}

\item{herbs}{standard data frame with herbs info}

\item{verbose}{A boolean flag to indicate console output}
}
\description{
This function takes row of a standard data frame of FIA, FR or ES and creates a list following the format of a forest object
}
