% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFN2forest.R
\encoding{UTF-8}
\name{IFN2forest}
\alias{IFN2forest}
\alias{IFN2forestlist}
\title{Extract forest from IFN data}
\usage{
IFN2forest(
  pies_mayores,
  matorral,
  ID,
  SpParams,
  IFNherbData = NULL,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  keepNumOrden = TRUE,
  verbose = TRUE
)

IFN2forestlist(
  pies_mayores,
  matorral,
  SpParams,
  IFNherbData = NULL,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  keepNumOrden = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pies_mayores}{A data frame with measured tree data (PCPiesMayores).}

\item{matorral}{A data frame with measured shrub data (PCMatorral).}

\item{ID}{A string with the ID of the plot to be extracted.}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{IFNherbData}{A data frame with cover and mean height of the herb layer.}

\item{setDefaults}{Initializes default values for missing fields in IFN data.}

\item{filterWrongRecords}{Filters wrong records (records with missing values, zero values or wrong growth forms)}

\item{keepNumOrden}{Keeps num orden as additional column (OrdenIfn2, OrdenIfn3, ...)}

\item{verbose}{A boolean flag to indicate console output.}
}
\value{
Function \code{IFN2forest} returns an object of class \code{\link{forest}}, whereas function \code{IFN2forestlist} returns a list of \code{\link{forest}} objects.
}
\description{
Creates a \code{\link{forest}} object from Spanish Forest Inventory (IFN) data (DGCN 2005).
}
\details{
IFN input data needs to be in a specific format.
\itemize{
  \item{
    For \code{pies_mayores}, the following columns are required:
    \itemize{
      \item{ID: Plot ID string}
      \item{Especie: String of tree IFN species code (will be mapped to medfate code).}
      \item{Dn1: Diameter at breast height (in mm), first measurement.}
      \item{Dn2: Diameter at breast height (in mm), second measurement.}
      \item{Ht: Total tree height (in m)}
    }
  }
  \item{
    For \code{matorral}, the following columns are required:
    \itemize{
       \item{ID: Plot ID string.}
       \item{Especie: String of shrub IFN species code (will be mapped to medfate code).}
       \item{Fcc: Numeric values of shrub cover (in \%).}
       \item{Hm: Average shrub height (in dm).}
    }
  }
}

Functions \code{IFN2forest} and \code{IFN2forestlist} call \code{\link{translateIFNSpeciesCodes}} internally
to translate IFN codes into medfate codes.
}
\examples{
data(SpParamsMED)
data(piesMayoresIFN2)
data(matorralIFN2)

# Builds an object 'forest' corresponding to one specific forest plot
f = IFN2forest(piesMayoresIFN2, matorralIFN2,
               ID = "081065", SpParams = SpParamsMED)
print(f)

# Builds a list whose elements are 'forest' objects
l = IFN2forestlist(piesMayoresIFN2, matorralIFN2, SpParamsMED)

# Plot codes are in list names
names(l)

# First forest object
l[[1]]
}
\references{
DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.
}
\seealso{
\code{\link{forest}}, \code{\link{translateIFNSpeciesCodes}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
