% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{IFN_species_mapping}
\alias{IFN_species_mapping}
\title{Species mapping for the Spanish NFI}
\description{
Data frame with the mapping of species codes from Spanish National Forest Inventory into medfate' species names
}
\keyword{data}
