% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSpParams.R
\encoding{UTF-8}
\name{initSpParams}
\alias{initSpParams}
\title{Initializes Species Parameter Table}
\usage{
initSpParams(
  sp_names,
  SpParamsDefinition,
  accepted_names = NULL,
  fill_taxonomic = TRUE,
  sort = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sp_names}{Vector of plant names, either taxon names or arbitrary species group names}

\item{SpParamsDefinition}{Data frame of species parameter definition from package medfate}

\item{accepted_names}{Vector of accepted taxon names of the same length of `sp_names`}

\item{fill_taxonomic}{Boolean flag to indicate that taxonomic information should be filled (retrieved from GBIF using package 'taxize')}

\item{sort}{Boolean flag to force sorting in ascending order by `Name`}

\item{verbose}{A boolean flag to indicate extra console output}
}
\value{
A data frame with empty species parameter values suitable for medfate simulations
}
\description{
Creates an empty plant parameter table for medfate, populating taxonomic information if desired.
}
\details{
Taxonomic information is retrieved using functions in package taxize and GBIF as data source.
}
\examples{
\dontrun{
# Load species parameter definition from medfate
data(SpParamsDefinition)

# Simple example with two species
sp_names <- c("Salvia rosmarinifolia", "Pinus contorta")
initSpParams(sp_names, SpParamsDefinition)

# Simple example with three species using synonyms and subspecies
sp_names <- c("Rosmarinus officinalis", "Pinus contorta", "Quercus ilex subsp. ilex")
accepted_names <- c("Salvia rosmarinifolia", "Pinus contorta", "Quercus ilex subsp. ilex")
initSpParams(sp_names, SpParamsDefinition, accepted_names)

# Initialisation with Spanish NFI species mapping
data(NFI_SP_mapping)
NFI_SP_mapping_red <- NFI_SP_mapping[!is.na(NFI_SP_mapping$originalName),]
SpParams <- initSpParams(sp_names = NFI_SP_mapping_red$originalName[1:20],
                        SpParamsDefinition = SpParamsDefinition,
                        accepted_names = NFI_SP_mapping_red$acceptedName[1:20],
                        verbose = TRUE)

}
}
\seealso{
\code{\link[medfate]{SpParamsMED}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
