% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medfateutils-package.R
\docType{package}
\name{medfateutils-package}
\alias{medfateutils-package}
\alias{medfateutils}
\title{medfateutils: Utility Functions for Package `medfate`}
\description{
Utility functions to facilitate defining inputs (soil, species or vegetation parameters) for simulations with
package medfate.
}
\seealso{
Useful links: \itemize{ \item{
\url{https://emf-creaf.github.io/medfateutils/index.html}} }
}
\author{
Maintainer: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
[ORCID](https://orcid.org/0000-0001-7132-2080)

Author: Víctor Granda
[ORCID](https://orcid.org/0000-0002-0469-1991)

Author: Adriana Tovar

Author: Antoine Cabon
[ORCID](https://orcid.org/0000-0001-6426-1726)

Author: Nicolas Martin-StPaul
[ORCID](https://orcid.org/0000-0001-7574-0108)

Author: Arsène Druel
[ORCID](https://orcid.org/0000-0002-3938-0085)
}
\keyword{internal}
