% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateTraits.R
\encoding{UTF-8}
\name{populateTraits}
\alias{populateTraits}
\title{Populate species parameters from trait data}
\usage{
populateTraits(
  SpParams,
  trait_table,
  trait_mapping,
  taxon_column = NULL,
  scalar_functions = NULL,
  replace_previous = FALSE,
  erase_previous = FALSE
)
}
\arguments{
\item{SpParams}{A data frame of medfate species parameters to be populated.}

\item{trait_table}{A data frame with species traits in columns and row.names corresponding to taxonomic entities.}

\item{trait_mapping}{A named string vector specifying which trait data column should used to populate each medfate param. Elements are data base columns and names are medfate params.}

\item{taxon_column}{A string identifying the column in \code{trait_table} that identifies taxa (normally species). If \code{taxon_column = NULL} then taxon names are taken from row.names.}

\item{scalar_functions}{A named list of scalar functions for traits needing transformation of units or scaling. Names are medfate params.}

\item{replace_previous}{A boolean flag to indicate that non-missing previous values should be replaced with new data}

\item{erase_previous}{A boolean flag to indicate that all previous values should be set to NA before populating with new data}
}
\value{
A modified data frame of medfate species parameters
}
\description{
Generic function to fill medfate species trait parameters from a data table of species traits
}
\seealso{
\code{\link{initSpParams}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
