% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateAllometries.R
\encoding{UTF-8}
\name{populateTreeAllometries}
\alias{populateTreeAllometries}
\title{Populate tree species allometries}
\usage{
populateTreeAllometries(
  SpParams,
  allom_table,
  allom_type = "foliarbiomass",
  erase_previous = FALSE
)
}
\arguments{
\item{SpParams}{A data frame of species parameters to be populated}

\item{allom_table}{A data frame of allometric parameters in columns and taxonomic entities (from species to group) as row names.}

\item{allom_type}{A string with the type of allometry to be filled, either "foliarbiomass", "crownwidth" or "crownratio".}

\item{erase_previous}{A boolean flag to indicate that values should be set to NA before populating with data}
}
\value{
A modified data frame of species parameters
}
\description{
Populates allometric coefficients for tree species from an input table, trying to use the taxonomic closest entity
}
\seealso{
\code{\link{initSpParams}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
