% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2forest_2.R
\name{sf2forest}
\alias{sf2forest}
\title{This function takes a standard  frame and generates an object with all plots and years}
\usage{
sf2forest(
  input_df,
  filterNA,
  filterDead,
  minDBH,
  filterminDBH,
  setDefaults,
  .verbose = TRUE
)
}
\arguments{
\item{input_df}{standard data frame}

\item{filterNA}{if TRUE records with NA for DBH or HT  are eliminated}

\item{filterDead}{if TRUE only records of live tree are conserved}

\item{minDBH}{min DBH value}

\item{filterminDBH}{if TRUE records are eliminated below minDBH}

\item{setDefaults}{set defaults param for roots}

\item{country}{country code "FR" "ES" or "US"}

\item{version}{NA as default, enter "ifn2" "ifn3" OR "ifn4" only for SPAIN}

\item{tree}{standard data frame with tree info}

\item{regen}{standard data frame with regen info}

\item{understory}{standard data frame with understory info}

\item{herbs}{standard data frame with herbs info}
}
\description{
This function takes a standard  frame and generates an object with all plots and years
}
