% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateIFNSpeciesCodes.R
\encoding{UTF-8}
\name{translateIFNSpeciesCodes}
\alias{translateIFNSpeciesCodes}
\title{Translate species codes}
\usage{
translateIFNSpeciesCodes(x, IFNcodes)
}
\arguments{
\item{x}{A vector of IFN species codes.}

\item{IFNcodes}{A string vector (of length equal to the number of rows in \code{SpParams} of the IFN species codes that correspond to the model species codification. Each string may contain different coma-separated codes in order to merge IFN species into a single model species.}
}
\value{
A string vector with the length of \code{x} and translated codes.
}
\description{
Translate IFN species codes to medfate SpIndex codes
}
\seealso{
\code{\link{IFN2forest}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
