% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRootingDepth.R
\encoding{UTF-8}
\name{estimateRootingDepth}
\alias{estimateRootingDepth}
\alias{estimateRootingDepthList}
\title{Estimate Rooting Depth Parameters}
\usage{
estimateRootingDepth(
  forest,
  soil,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

estimateRootingDepthList(
  forestlist,
  soillist,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)
}
\arguments{
\item{forest}{An object of class \code{\link{forest}}.}

\item{soil}{An object of class \code{\link{soil}} or a data frame with soil parameters.}

\item{PET_summer}{Average (climatic) potential evapotranspiration (mm) corresponding to summer months.}

\item{P_summer}{Average (climatic) precipitation (mm) corresponding to summer months.}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{fillMissing}{A boolean flag to indicate that missing estimates should be filled with mean values.}

\item{forestlist}{A list of objects of class \code{\link{forest}}.}

\item{soillist}{A list with objects of class \code{\link{soil}} or data frames with soil parameters.}
}
\value{
Function \code{estimateRootingDepth} returns object of class \code{\link{forest}} with modified values for \code{Z50} and \code{Z95}. Function \code{estimateRootingDepthList} returns a list of forest objects with modified root distribution parameters.
}
\description{
Estimates rooting depth parameters for forest objects following Cabon et al. (2018).
}
\details{
If \code{PET_summer} or \code{P_summer} are of length one, their elements are repeated for all elements in \code{forestlist}. Similarly, if \code{soillist} is an object of class \code{\link{soil}} or a data frame with soil parameters, the same soil parameters are repeated for all elements in \code{forestlist}.
}
\examples{
data(SpParamsMED)
data(example_treedata_ifn)
data(example_shrubdata_ifn)

# Builds an object 'forest' corresponding to one specific forest plot
f1 = IFN2forest(example_treedata_ifn, example_shrubdata_ifn,
                ID = "81065", SpParams = SpParamsMED)
print(f1)

soilpar = defaultSoilParams()
f2 = estimateRootingDepth(f1, soilpar, 300, 100, SpParamsMED)
print(f2)

l1 = IFN2forestlist(example_treedata_ifn, example_shrubdata_ifn, SpParamsMED)
l2 = estimateRootingDepthList(l1, soilpar, 300, 100, SpParamsMED)
}
\references{
Cabon, A., Martinez-Vilalta, J., de Aragon, J.M., Poyatos, R., De Caceres, M., 2018. Applying the eco-hydrological equilibrium hypothesis to model root distribution in water-limited forests. Ecohydrology 11, e2015. https://doi.org/10.1002/eco.2015
}
\seealso{
\code{\link[medfate]{forest}}, \code{\link{translateIFNSpeciesCodes}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
