% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFN2forest.R
\encoding{UTF-8}
\name{IFN2forest}
\alias{IFN2forest}
\alias{IFN2forestlist}
\title{Extract forest from IFN data}
\usage{
IFN2forest(
  IFNtreeData,
  IFNshrubData,
  ID,
  SpParams,
  IFNherbData = NULL,
  patchsize = 10000,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  keepNumOrden = TRUE,
  verbose = TRUE
)

IFN2forestlist(
  IFNtreeData,
  IFNshrubData,
  SpParams,
  IFNherbData = NULL,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  keepNumOrden = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{IFNtreeData}{A data frame with measured tree data.}

\item{IFNshrubData}{A data frame with measured shrub data.}

\item{ID}{A string with the ID of the plot to be extracted.}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{IFNherbData}{A data frame with cover and mean height of the herb layer.}

\item{patchsize}{The area of the forest stand, in square meters.}

\item{setDefaults}{Initializes default values for missing fields in IFN data.}

\item{filterWrongRecords}{Filters wrong records (records with missing values, zero values or wrong growth forms)}

\item{keepNumOrden}{Keeps num orden as additional column (OIF2, OIF3, ...)}

\item{verbose}{A boolean flag to indicate console output.}
}
\value{
Function \code{IFN2forest} returns an object of class \code{\link{forest}}, whereas function \code{IFN2forestlist} returns a list of \code{\link{forest}} objects.
}
\description{
Creates a \code{\link{forest}} object from Spanish Forest Inventory (IFN) data (DGCN 2005).
}
\details{
IFN input data needs to be in a specific format, see package \code{IFNread}. Functions \code{IFN2forest} and \code{IFN2forestlist} call \code{\link{translateIFNSpeciesCodes}} internally.
}
\examples{
data(SpParamsMED)
data(exampletreedata)
data(exampleshrubdata)

# Builds an object 'forest' corresponding to one specific forest plot
f = IFN2forest(exampletreedata, exampleshrubdata,
               ID = "81065", SpParams = SpParamsMED)
print(f)

# Builds a list whose elements are 'forest' objects
l = IFN2forestlist(exampletreedata, exampleshrubdata, SpParamsMED)
names(l)
l[[1]]
}
\references{
DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.
}
\seealso{
\code{\link{forest}}, \code{\link{translateIFNSpeciesCodes}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
