% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFN2forest.R
\encoding{UTF-8}
\name{IFN2forest}
\alias{IFN2forest}
\title{Extract forest from IFN data}
\usage{
IFN2forest(
  pies_mayores,
  SpParams,
  pies_menores = NULL,
  regenera = NULL,
  matorral = NULL,
  herb_data = NULL,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  keepNumOrden = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pies_mayores}{A data frame with measured tree data (PCPiesMayores).}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{pies_menores}{A data frame with measured regeneration tree data (PCPiesMenores in IFN2).}

\item{regenera}{A data frame with measured regeneration tree data (PCRegenera in IFN3 and IFN4).}

\item{matorral}{A data frame with measured shrub data (PCMatorral).}

\item{herb_data}{A data frame with cover and mean height of the herb layer for each forest plot.}

\item{setDefaults}{Initializes default values for missing fields in IFN data.}

\item{filterWrongRecords}{Filters wrong records (records with missing values, zero values or wrong growth forms).
This should normally result in the removal of dead/cut trees.}

\item{keepNumOrden}{Keeps num orden as additional column (OrdenIf2, OrdenIf3 or OrdenIf4) to identify trees.}

\item{verbose}{A boolean flag to indicate console output.}
}
\value{
A list of \code{\link{forest}} objects.
}
\description{
Creates a list of \code{\link{forest}} objects from Spanish Forest Inventory (IFN) data in its
second (IFN2), third (IFN3; DGCN 2005) and fourth (IFN4) editions.
}
\details{
IFN input data needs to be in a specific format, following IFN specifications.
\itemize{
  \item{
    For \code{pies_mayores}, the following columns are required:
    \itemize{
      \item{ID: Plot ID string}
      \item{Especie: String of tree IFN species code (will be mapped to medfate code).}
      \item{Dn1: Diameter at breast height (in mm), first measurement.}
      \item{Dn2: Diameter at breast height (in mm), second measurement.}
      \item{Ht: Total tree height (in m)}
    }
  }
  \item{
    For \code{pies_menores}, the following columns are required:
    \itemize{
      \item{ID: Plot ID string}
      \item{Especie: String of tree IFN species code (will be mapped to medfate code).}
      \item{Numero: Number of stems for 2.5-7.5 cm DBH class.}
      \item{Hm: Average height (in dm).}
      \item{Regena: Regeneration code ('0', 1', '2', '3').}
    }
  }
  \item{
    For \code{regenera}, the following columns are required:
    \itemize{
      \item{ID: Plot ID string}
      \item{Especie: String of tree IFN species code (will be mapped to medfate code).}
      \item{CatDes: Regeneration category ( 1', '2', '3' or '4').}
      \item{Densidad: Density category ( 1', '2', or '3').}
      \item{NumPies: Number of stems for 2.5-7.5 cm DBH class.}
      \item{Hm: Average height (in dm).}
    }
  }
  \item{
    For \code{matorral}, the following columns are required:
    \itemize{
       \item{ID: Plot ID string.}
       \item{Especie: String of shrub IFN species code (will be mapped to medfate code).}
       \item{Fcc: Numeric values of shrub cover (in \%).}
       \item{Hm: Average shrub height (in dm).}
    }
  }
}

Function \code{IFN2forest} call \code{\link{translateIFNSpeciesCodes}} internally
to translate IFN codes into medfate codes.
}
\examples{

# Medfate species parameters
data(SpParamsMED)

# Builds from IFN2 data a list whose elements are 'forest' objects
data(piesMenoresIFN2)
data(piesMayoresIFN2)
data(matorralIFN2)
l <- IFN2forest(piesMayoresIFN2, SpParamsMED,
               matorral = matorralIFN2, pies_menores = piesMenoresIFN2)

# Plot codes are in list names
names(l)


# Builds from IFN3 data a list whose elements are 'forest' objects
data(piesMayoresIFN3)
data(regeneraIFN3)
data(matorralIFN3)
l <- IFN2forest(piesMayoresIFN3, SpParamsMED,
               matorral = matorralIFN3, regenera = regeneraIFN3)

}
\references{
DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.
}
\seealso{
\code{\link{forest}}, \code{\link{translateIFNSpeciesCodes}} \code{\link{example_datasets}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
