% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSpParams.R
\encoding{UTF-8}
\name{initSpParams}
\alias{initSpParams}
\title{Initializes Species Parameter Table}
\usage{
initSpParams(
  sp_names,
  SpParamsDefinition,
  fill_taxonomic = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sp_names}{Vector of species names, taxon names or arbitrary species group names}

\item{SpParamsDefinition}{Data frame of species parameter definition from package 'medfate'}

\item{fill_taxonomic}{Boolean flag to indicate that taxonomic information should be filled (retrieved from GBIF using package 'taxize')}

\item{verbose}{A boolean flag to indicate extra console output}
}
\value{
A data frame with empty species parameter values suitable for medfate simulations
}
\description{
Creates an empty species parameter table for medfate, populating taxonomic information if desired.
}
\details{
Taxonomic information is retrieved using functions in package taxize and GBIF as data source.
}
\examples{
\dontrun{
# Load species parameter definition from medfate
data(SpParamsDefinition)

# Simple example with two species
sp_names = c("Salvia rosmarinifolia", "Pinus contorta")
initSpParams(sp_names, SpParamsDefinition)

# Initialisation with IFN species mapping
data(IFN_species_mapping)
initSpParams(IFN_species_mapping$Name, SpParamsDefinition, verbose = TRUE)

}
}
\seealso{
\code{\link[medfate]{SpParamsMED}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
