% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSpParams.R
\encoding{UTF-8}
\name{initSpParams}
\alias{initSpParams}
\title{Initializes Species Parameter Table}
\usage{
initSpParams(
  sp_codes,
  sp_names,
  group_codes = NULL,
  group_names = NULL,
  fill_taxonomic = TRUE
)
}
\arguments{
\item{sp_codes}{Vector of IFN species codes}

\item{sp_names}{Vector of IFN species names}

\item{group_codes}{List of species group codes (strings where species of the same group are separated by "/", e.g. "code1/code2/code3")}

\item{group_names}{Vector of group names of the same length as \code{group_codes}}

\item{fill_taxonomic}{Boolean flag to indicate that taxonomic information should be filled (retrieved from GBIF using package 'taxize')}
}
\value{
A data frame with empty species parameter values suitable for medfate simulations
}
\description{
Creates an empty species parameter table for medfate, populating taxonomic information if desired.
}
\details{
Taxonomic information is retrieved using functions in package taxize and GBIF as data source.
}
\examples{

}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
