% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifySoilParams.R
\name{modifySoilParams}
\alias{modifySoilParams}
\title{Modifies soil properties estimated from SoilGrids}
\usage{
modifySoilParams(soildata, soildepth, surfacerock = NULL)
}
\arguments{
\item{soildata}{A data frame of soil physical parameters or a list of data frames}

\item{soildepth}{An estimate of soil depth (in mm) or a vector of soil depths if `soildata` is a list}

\item{surfacerock}{An estimate of surface rock content (in percent) or a vector of such values if `soildata` is a list}
}
\value{
A data frame of modified physical parameters or a list of modified data frames
}
\description{
Modifies soil properties estimated from SoilGrids
}
\examples{

 \dontrun{
  library(sp)
  coords_sp <- SpatialPoints(cbind(long = -5.6333, lat = 42.6667),
                                 CRS(SRS_string = "EPSG:4326"))
  foo <- soilgridsParams(coords_sp, widths = c(300, 700, 1000, 2000))
  foo
  foo_mod <- modifySoilParams(foo, 900, 20) # 90 cm depth, 20 \% rocks in the surface
  foo_mod
 }
}
