% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spwb_rockOptimization.R
\name{spwb_rockOptimization}
\alias{spwb_rockOptimization}
\title{Optimization of rock fragment content}
\usage{
spwb_rockOptimization(
  x,
  meteo,
  PLC90_target = 12,
  PLC90_tol = 0.5,
  max_simu = 7,
  model_varLim = 10,
  max_rocks = 99,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link[medfate]{spwbInput}}.}

\item{meteo}{A data frame with daily meteorological data series (see \code{\link[medfate]{spwb}}).}

\item{PLC90_target}{Stem PLC target (quantile 90).}

\item{PLC90_tol}{Limit of the PLC target tolerance, only in some conditions.}

\item{max_simu}{Maximum of simulation authorized before stopping.}

\item{model_varLim}{Limit of the soil extractable water (SEW) variation of the model accepted.}

\item{max_rocks}{Maximum content in coarse fragments allowed for any soil layer.}

\item{verbose}{A logical value. Print the internal messages of the function?}

\item{...}{Additional parameters to function \code{\link[medfate]{spwb}}.}
}
\value{
Function \code{spwb_rockOptimization} returns a list containing the estimated rock fragment content,
the corresponding soil extractable water and a dataframe of simulation results needed for the estimation.
}
\description{
Function \code{spwb_rockOptimization} finds optimum rock fragment content in the soil
corresponding to given vegetation, weather and target percent loss
of conductance (PLC), following the method described in Druel et al. (2023).
}
\details{
The function performs a model inversion based on an ecohydrological assumption,
consisting in that forest leaf area index is in equilibrium with a low embolism
rate under normal conditions. This is translated in that the 90\% interannual quantile of
the maximum annual percent loss of conductance (PLC), averaged over plant cohorts,
should be close to a target PLC value (by default 12\%).
}
\examples{
\donttest{
#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params
examplesoil <- soil(defaultSoilParams(2))

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- spwbInput(exampleforest,examplesoil, SpParamsMED, control)

#Rock fragment content optimization
spwb_rockOptimization(x, meteo = examplemeteo,
                      elevation = 100, latitude = 41.82592)
}
}
\references{
Druel, A., Martins, N., Cochard, H., De Caceres, M., Delzon, S., Mencuccini, M., Torres-Ruiz, J., and Ruffault, J.: European forest vulnerability to hydraulic failure: an ecohydrological approach, EGU General Assembly 2023, Vienna, Austria, 24–28 Apr 2023, EGU23-17068, https://doi.org/10.5194/egusphere-egu23-17068, 2023.
}
\seealso{
\code{\link[medfate]{spwb}}, \code{\link[medfate]{soil}}, \code{\link{spwb_ldrOptimization}}
}
\author{
\enc{Arsène}{Arsene} Druel, URFM-INRAE

Nicolas Martin-StPaul, URFM-INRAE

Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
